requirements.txt
python-docx==1.2.0
lxml==6.0.1
beautifulsoup4==4.13.5
mammoth==1.10.0